# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "GuestCommentIn": {
        "properties": {
          "body": {
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "task_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          }
        },
        "required": [
          "name",
          "body"
        ],
        "title": "GuestCommentIn",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Planora",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/public/{org_slug}/{project_slug}": {
      "get": {
        "description": "The same layout as on the working screen, but without internal notes and\nwithout assignees.",
        "operationId": "public_project_api_public__org_slug___project_slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "org_slug",
            "required": true,
            "schema": {
              "title": "Org Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_slug",
            "required": true,
            "schema": {
              "title": "Project Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "s",
            "required": false,
            "schema": {
              "default": "",
              "title": "S",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Public Project",
        "tags": [
          "public"
        ]
      }
    },
    "/api/public/{org_slug}/{project_slug}/comments": {
      "get": {
        "description": "The feed is visible even when comments are turned off.\n\nTurned-off comments are a ban on writing, not an order to hide what has\nalready been said: a conversation the client saw yesterday must not vanish at\nthe flick of a toggle.\n\nA guest does not see internal remarks: that is the team's conversation\n\"aside\", not part of the public page.",
        "operationId": "public_comments_api_public__org_slug___project_slug__comments_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_slug",
            "required": true,
            "schema": {
              "title": "Org Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_slug",
            "required": true,
            "schema": {
              "title": "Project Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "task_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          },
          {
            "in": "query",
            "name": "s",
            "required": false,
            "schema": {
              "default": "",
              "title": "S",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Public Comments",
        "tags": [
          "public"
        ]
      },
      "post": {
        "operationId": "add_public_comment_api_public__org_slug___project_slug__comments_post",
        "parameters": [
          {
            "in": "path",
            "name": "org_slug",
            "required": true,
            "schema": {
              "title": "Org Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_slug",
            "required": true,
            "schema": {
              "title": "Project Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "s",
            "required": false,
            "schema": {
              "default": "",
              "title": "S",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestCommentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Public Comment",
        "tags": [
          "public"
        ]
      }
    },
    "/api/public/{org_slug}/{project_slug}/comments/counts": {
      "get": {
        "description": "The remark counter on the rows of the public chart — without internal ones.\n\nThe same filter as on the feed above: a guest does not see internal remarks,\nand the number next to a task must not let slip what is not in their feed at\nall.",
        "operationId": "public_comment_counts_api_public__org_slug___project_slug__comments_counts_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_slug",
            "required": true,
            "schema": {
              "title": "Org Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_slug",
            "required": true,
            "schema": {
              "title": "Project Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "s",
            "required": false,
            "schema": {
              "default": "",
              "title": "S",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Public Comment Counts",
        "tags": [
          "public"
        ]
      }
    },
    "/api/public/{org_slug}/{project_slug}/export.pdf": {
      "get": {
        "operationId": "public_export_pdf_api_public__org_slug___project_slug__export_pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_slug",
            "required": true,
            "schema": {
              "title": "Org Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_slug",
            "required": true,
            "schema": {
              "title": "Project Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "s",
            "required": false,
            "schema": {
              "default": "",
              "title": "S",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/pdf": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export the project through a public link as a PDF document",
        "tags": [
          "public"
        ]
      }
    },
    "/api/public/{org_slug}/{project_slug}/export.xlsx": {
      "get": {
        "operationId": "public_export_xlsx_api_public__org_slug___project_slug__export_xlsx_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_slug",
            "required": true,
            "schema": {
              "title": "Org Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_slug",
            "required": true,
            "schema": {
              "title": "Project Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "s",
            "required": false,
            "schema": {
              "default": "",
              "title": "S",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export the project through a public link as an Excel workbook",
        "tags": [
          "public"
        ]
      }
    }
  }
}